refactor: routing for startpage and embed
