+++ snowpack.config.js
@@ -18,8 +18,5 @@ module.exports = {
       resolve: true,
     },
   },
-  devOptions: {
-    openUrl: '?url=http%3A%2F%2Flocalhost%3A8080%2Fexample',
-  },
   exclude: ['public/**/*'],
 };
