--- src/app.js
@@ -1,5 +1,0 @@
-import { initNavigation } from './navigation/navigation';
-import { initSandbox } from './main';
-
-initNavigation();
-initSandbox();
