+++ src/embed/constants.js
