+++ src/embed/navigation/navigation.css
