+++ src/embed/styles.css
@@ -1,0 +1,60 @@
+@import 'prismjs/themes/prism.css';
+@import './navigation/navigation.css';
+
+.body {
+  margin: 0;
+  padding: 0;
+  border: 4px solid #eee;
+  height: calc(100vh - 58px);
+}
+
+.container {
+  display: flex;
+  height: 100%;
+}
+
+.result {
+  width: 100%;
+  border: none;
+  display: none;
+}
+
+.previews {
+  display: flex;
+  flex-direction: column;
+  display: none;
+  width: 100%;
+}
+
+.preview {
+  overflow: auto;
+  margin: 0 !important;
+  height: 100%;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.preview > code {
+  margin: 0 !important;
+}
+
+.preview-html,
+.preview-css,
+.preview-js {
+  display: none;
+}
+
+.preview-html.active,
+.preview-css.active,
+.preview-js.active {
+  display: block;
+  width: 100%;
+}
+
+.previews.active {
+  display: block;
+}
+
+.result.active {
+  display: block;
+}
