+++ src/styles.css
@@ -1,59 +1,73 @@
-@import './navigation/navigation.css';
-
-.body {
-  margin: 0;
-  padding: 0;
-  border: 4px solid #eee;
-  height: calc(100vh - 58px);
-}
-
-.container {
-  display: flex;
-  height: 100%;
-}
-
-.result {
-  width: 100%;
+/* button */
+button {
+  color: #fff;
   border: none;
-  display: none;
+  padding: 0.5rem 1.5rem;
+  cursor: pointer;
+  font-weight: 700;
+  background-color: var(--grey-900);
 }
 
-.previews {
+button:hover {
+  background-color: var(--grey-700);
+}
+
+button:disabled {
+  background-color: var(--grey-300);
+  pointer-events: none;
+}
+
+code {
+  font-family: 'Source Code Pro', Monaco, 'Andale Mono', 'Ubuntu Mono',
+    monospace;
+  font-style: normal;
+  background-color: var(--grey-300);
+}
+
+pre code {
+  display: block;
+  overflow-x: auto;
+  font-weight: normal;
+  background-color: transparent;
+}
+
+label {
+  font-weight: 700;
+}
+
+input,
+textarea {
+  border: 4px solid var(--grey-900);
+  padding: 0.5rem;
+  margin-top: 0.5rem;
+}
+
+input:disabled,
+textarea:disabled {
+  border: 4px solid var(--grey-300);
+}
+
+.documentation-hint {
+  background-color: var(--blue-300);
+  border: 4px solid var(--blue-500);
+  padding: 0.8rem;
+}
+
+.form__wrapper {
   display: flex;
   flex-direction: column;
-  display: none;
-  width: 100%;
 }
 
-.preview {
-  overflow: auto;
-  margin: 0 !important;
-  height: 100%;
-  width: 100%;
-  box-sizing: border-box;
+.about-images {
+  display: flex;
 }
 
-.preview > code {
-  margin: 0 !important;
+* + figure {
+  margin-top: 0;
+  margin-left: 1rem;
 }
 
-.preview-html,
-.preview-css,
-.preview-js {
-  display: none;
-}
-
-.preview-html.active,
-.preview-css.active,
-.preview-js.active {
-  display: block;
-  width: 100%;
-}
-
-.previews.active {
-  display: block;
-}
-
-.result.active {
-  display: block;
+div figure figcaption {
+  margin-top: 0;
+  text-align: center;
 }
