--- website/styles.css
@@ -1,73 +1,0 @@
-/* button */
-button {
-  color: #fff;
-  border: none;
-  padding: 0.5rem 1.5rem;
-  cursor: pointer;
-  font-weight: 700;
-  background-color: var(--grey-900);
-}
-
-button:hover {
-  background-color: var(--grey-700);
-}
-
-button:disabled {
-  background-color: var(--grey-300);
-  pointer-events: none;
-}
-
-code {
-  font-family: 'Source Code Pro', Monaco, 'Andale Mono', 'Ubuntu Mono',
-    monospace;
-  font-style: normal;
-  background-color: var(--grey-300);
-}
-
-pre code {
-  display: block;
-  overflow-x: auto;
-  font-weight: normal;
-  background-color: transparent;
-}
-
-label {
-  font-weight: 700;
-}
-
-input,
-textarea {
-  border: 4px solid var(--grey-900);
-  padding: 0.5rem;
-  margin-top: 0.5rem;
-}
-
-input:disabled,
-textarea:disabled {
-  border: 4px solid var(--grey-300);
-}
-
-.documentation-hint {
-  background-color: var(--blue-300);
-  border: 4px solid var(--blue-500);
-  padding: 0.8rem;
-}
-
-.form__wrapper {
-  display: flex;
-  flex-direction: column;
-}
-
-.about-images {
-  display: flex;
-}
-
-* + figure {
-  margin-top: 0;
-  margin-left: 1rem;
-}
-
-div figure figcaption {
-  margin-top: 0;
-  text-align: center;
-}
